improve(build/luamake): rework dependencies
